+++ lib/controller/component/common/get-composite.js
@@ -4,7 +4,7 @@ const {
   getServiceSchema
 } = require('~/fb-runner-node/service-data/service-data')
 
-const getComponentController = require('./get-component-controller')
+const getComponentController = require('~/fb-runner-node/controller/component/get-controller')
 
 module.exports = function getComponentComposite (componentInstance) {
   if (componentInstance) {
