+++ lib/controller/component/common/get-composite.unit.spec.js
@@ -11,9 +11,9 @@ const getComponentControllerStub = stub().returns({
   getComposite: getCompositeStub
 })
 
-const getComponentComposite = proxyquire('./get-component-composite', {
+const getComponentComposite = proxyquire('./get-composite', {
   '~/fb-runner-node/service-data/service-data': getServiceSchemaStub,
-  './get-component-controller': getComponentControllerStub
+  '~/fb-runner-node/controller/component/get-controller': getComponentControllerStub
 })
 
 getServiceSchemaStub.callsFake((_type) => {
