+++ lib/controller/component/get-controller.js
