+++ lib/controller/component/type/answers/answers.common.js
@@ -3,7 +3,7 @@ require('@ministryofjustice/module-alias/register-module')(module)
 const jsonPath = require('jsonpath')
 const striptags = require('striptags')
 
-const getComponentController = require('~/fb-runner-node/controller/component/get-component-controller')
+const getComponentController = require('~/fb-runner-node/controller/component/get-controller')
 
 const {
   getUrl,
