+++ lib/controller/component/type/answers/answers.common.unit.spec.js
@@ -82,7 +82,7 @@ const {
   '~/fb-runner-node/format/format': {
     format: formatStub
   },
-  '~/fb-runner-node/controller/component/get-component-controller': getComponentControllerStub
+  '~/fb-runner-node/controller/component/get-controller': getComponentControllerStub
 })
 
 test('The current page has the same `_id` as the page instance', (t) => {
