+++ lib/controller/component/type/answers/answers.page-summary.js
@@ -22,7 +22,7 @@ const {
   setRepeatable
 } = require('~/fb-runner-node/page/page')
 
-const getComponentController = require('~/fb-runner-node/controller/component/get-component-controller')
+const getComponentController = require('~/fb-runner-node/controller/component/get-controller')
 
 const {
   format
