+++ lib/middleware/routes-metadata/routes-metadata.js
@@ -53,8 +53,8 @@ const {
   kludgeUpdates
 } = require('~/fb-runner-node/page/page')
 
-const getPageController = require('~/fb-runner-node/controller/page/get-page-controller')
-const getComponentController = require('~/fb-runner-node/controller/component/get-component-controller')
+const getPageController = require('~/fb-runner-node/controller/page/get-controller')
+const getComponentController = require('~/fb-runner-node/controller/component/get-controller')
 
 const setFileUploads = require('~/fb-runner-node/page/set-fileuploads/set-fileuploads')
 const setFileUploadControlsMaxSize = require('~/fb-runner-node/page/set-fileuploads/set-fileupload-controls-max-size')
