+++ lib/page/get-display-value/get-display-value.js
@@ -1,6 +1,6 @@
 require('@ministryofjustice/module-alias/register-module')(module)
 
-const getComponentController = require('~/fb-runner-node/controller/component/get-component-controller')
+const getComponentController = require('~/fb-runner-node/controller/component/get-controller')
 
 function getDisplayValue (nameInstance, userData) {
   return getComponentController(nameInstance)
