+++ lib/page/get-display-value/get-display-value.unit.spec.js
@@ -22,7 +22,7 @@ const userData = {
 
 const getDisplayValue = proxyquire('./get-display-value', {
   bytes: bytesStub,
-  '~/fb-runner-node/controller/component/get-component-controller': getComponentControllerStub,
+  '~/fb-runner-node/controller/component/get-controller': getComponentControllerStub,
   '~/fb-runner-node/service-data/service-data': serviceData,
   '~/fb-runner-node/update-control-names/update-control-names': updateControlNames,
   '~/fb-runner-node/format/format': format
