+++ lib/page/process-input/process-control-composite.js
@@ -1,6 +1,6 @@
 require('@ministryofjustice/module-alias/register-module')(module)
 
-const getComponentController = require('~/fb-runner-node/controller/component/get-component-controller')
+const getComponentController = require('~/fb-runner-node/controller/component/get-controller')
 
 module.exports = function processControlComposite (pageInstance, userData, nameInstance, options) {
   const {
