+++ lib/page/process-input/process-control-composite.unit.spec.js
@@ -12,7 +12,7 @@ getComponentControllerStub.returns({
 })
 
 const processControlComposite = proxyquire('./process-control-composite', {
-  '~/fb-runner-node/controller/component/get-component-controller': getComponentControllerStub
+  '~/fb-runner-node/controller/component/get-controller': getComponentControllerStub
 })
 
 const getBodyInputStub = stub()
