+++ lib/page/process-input/process-control.js
@@ -1,6 +1,6 @@
 require('@ministryofjustice/module-alias/register-module')(module)
 
-const getComponentComposite = require('~/fb-runner-node/controller/component/get-component-composite')
+const getComponentComposite = require('~/fb-runner-node/controller/component/common/get-composite')
 
 const skipControlValidation = require('./skip-control-validation')
 const skipControlProcessing = require('./skip-control-processing')
