+++ lib/page/process-input/process-control.unit.spec.js
@@ -14,7 +14,7 @@ const updateControlValueStub = stub()
 const updateControlDisplayValueStub = stub()
 
 const processControl = proxyquire('./process-control', {
-  '~/fb-runner-node/controller/component/get-component-composite': getComponentCompositeStub,
+  '~/fb-runner-node/controller/component/common/get-composite': getComponentCompositeStub,
   './skip-control-validation': skipControlValidationStub,
   './skip-control-processing': skipControlProcessingStub,
   './process-control-composite': processControlCompositeStub,
