+++ lib/page/set-composite/set-composite.js
@@ -3,7 +3,7 @@ require('@ministryofjustice/module-alias/register-module')(module)
 const jp = require('jsonpath')
 
 const {getString} = require('~/fb-runner-node/service-data/service-data')
-const getComponentComposite = require('~/fb-runner-node/controller/component/get-component-composite')
+const getComponentComposite = require('~/fb-runner-node/controller/component/common/get-composite')
 
 const setComposite = (pageInstance, userData) => {
   const {getAllData, getUserDataProperty} = userData
