+++ lib/page/update-control-names/update-control-names.js
@@ -4,7 +4,7 @@ const jp = require('jsonpath')
 
 const {instanceHasError} = require('~/fb-runner-node/page/validate-input/validate-input')
 const redact = require('~/fb-runner-node/redact/redact')
-const getComponentComposite = require('~/fb-runner-node/controller/component/get-component-composite')
+const getComponentComposite = require('~/fb-runner-node/controller/component/common/get-composite')
 
 const getRedactedValue = (nameInstance, userData, skipRedaction, scope) => {
   let value = userData.getUserDataProperty(nameInstance.name, undefined, scope)
