+++ lib/page/validate-input/validate-input.js
@@ -17,7 +17,7 @@ const {
   validateInstanceValue
 } = require('./validate-value/validate-value')
 
-const getPageController = require('~/fb-runner-node/controller/page/get-page-controller')
+const getPageController = require('~/fb-runner-node/controller/page/get-controller')
 
 const flattenDeep = (arr) => arr.reduce((acc, val) => Array.isArray(val) ? acc.concat(flattenDeep(val)) : acc.concat(val), [])
