+++ lib/page/validate-input/validate-input.unit.spec.js
@@ -42,7 +42,7 @@ function initStubs () {
     '~/fb-runner-node/page/set-errors/set-errors': stubs.setErrorsStub,
     './validate-required/validate-required': stubs.validateRequiredStub,
     './validate-value/validate-value': stubs.validateValueStub,
-    '~/fb-runner-node/controller/page/get-page-controller': stubs.getPageControllerStub
+    '~/fb-runner-node/controller/page/get-controller': stubs.getPageControllerStub
   }
 
   for (const [stubName, stubValue] of Object.entries(stubConfig)) {
