+++ lib/page/validate-input/validate-required/validate-required.js
@@ -5,7 +5,7 @@ const {deepClone} = require('@ministryofjustice/fb-utils-node')
 
 const {getServiceSchema} = require('~/fb-runner-node/service-data/service-data')
 const {initAddError} = require('~/fb-runner-node/page/set-errors/set-errors')
-const getComponentComposite = require('~/fb-runner-node/controller/component/get-component-composite')
+const getComponentComposite = require('~/fb-runner-node/controller/component/common/get-composite')
 
 const isRequired = (instance, userData) => {
   // TODO: fix checkbox specification
