+++ lib/page/validate-input/validate-required/validate-required.unit.spec.js
@@ -14,7 +14,7 @@ const {
   validateRequired,
   validateInstanceRequired
 } = proxyquire('./validate-required', {
-  '~/fb-runner-node/controller/component/get-component-composite': getComponentCompositeStub
+  '~/fb-runner-node/controller/component/common/get-composite': getComponentCompositeStub
 })
 const {getUserDataMethods} = require('~/fb-runner-node/middleware/user-data/user-data')
 const getUserData = (input) => getUserDataMethods({input})
@@ -437,7 +437,7 @@ test('When checking whether a composite instance has an error', t => {
   t.end()
 })
 
-test('When checking whether a composite instance with a conttroller has an error', t => {
+test('When checking whether a composite instance with a controller has an error', t => {
   getComponentCompositeStub.returns(['x', 'y'])
   const getServiceSchemaStub = stub(validateRequired, 'getServiceSchema')
   getServiceSchemaStub.callsFake(() => {
